Edited some Entities and Created Service Logic and Controller Endpoints

+++ src/Services/MelonBookshelfApi/MelonBookchelfApi.Infrastructure/Data/Models/Resource.cs
@@ -35,7 +35,7 @@ namespace MelonBookchelfApi.Infrastructure.Data.Models
 
         public DateTime DateAdded { get; set; }
 
-        public decimal PurchasePrice { get; set; }
+        public double PurchasePrice { get; set; }
 
         public string? InvoiceAttachment { get; set; }
